+++ config/response_builder.php
@@ -41,7 +41,7 @@ return [
 	| readable form, that's why this mapping is needed. ResponseBuilder uses
 	| standard Laravel's Lang
 	|
-	|    ErrorCode::SOMETHING => 'api.something',
+	|    ApiCode::SOMETHING => 'api.something',
 	|
 	| See README if you want to provide own messages for built-in codes too.
 	|
@@ -112,19 +112,19 @@ return [
 		// you will still know which module thrown this exception
 		'exception' => [
 //			'http_not_found' => [
-//				'code'      => \App\ApiCodes::HTTP_NOT_FOUND,
+//				'code'      => \App\ApiCode::HTTP_NOT_FOUND,
 //				'http_code' => Symfony\Component\HttpFoundation\Response\::HTTP_BAD_REQUEST,
 //			],
 //			'http_service_unavailable' => [
-//				'code'      => \App\ApiCodes::HTTP_SERVICE_UNAVAILABLE,
+//				'code'      => \App\ApiCode::HTTP_SERVICE_UNAVAILABLE,
 //				'http_code' => Symfony\Component\HttpFoundation\Response\::HTTP_BAD_REQUEST,
 //			],
 //			'http_exception' => [
-//				'code'      => \App\ApiCodes::HTTP_EXCEPTION,
+//				'code'      => \App\ApiCode::HTTP_EXCEPTION,
 //				'http_code' => Symfony\Component\HttpFoundation\Response\::HTTP_BAD_REQUEST,
 //			],
 //			'uncaught_exception' => [
-//				'code'      => \App\ApiCodes::UNCAUGHT_EXCEPTION,
+//				'code'      => \App\ApiCode::UNCAUGHT_EXCEPTION,
 //				'http_code' => Symfony\Component\HttpFoundation\Response\::HTTP_INTERNAL_SERVER_ERROR,
 //			],
 		],