+++ tests/ErrorTest.php
@@ -15,24 +15,25 @@ use Symfony\Component\HttpFoundation\Response as HttpResponse;
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
  * @link      https://github.com/MarcinOrlowski/laravel-api-response-builder
  */
-
 class ErrorTest extends Base\ResponseBuilderTestCaseBase
 {
 
 	/**
 	 * Check success()
+	 *
+	 * @return void
 	 */
 	public function testError()
 	{
 		// GIVEN random error code
-		$error_code = $this->random_api_code;
+		$api_code = $this->random_api_code;
 
 		// WHEN we report error
-		$this->response = ResponseBuilder::error($error_code);
+		$this->response = ResponseBuilder::error($api_code);
 
 		// THEN returned message contains given error code and mapped message
-		$j = $this->getResponseErrorObject($error_code);
-		$this->assertEquals($this->error_message_map[ $error_code ], $j->message);
+		$j = $this->getResponseErrorObject($api_code);
+		$this->assertEquals($this->error_message_map[ $api_code ], $j->message);
 
 		// AND no data
 		$this->assertNull($j->data);
@@ -40,6 +41,8 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 
 	/**
 	 * Tests error() with various http codes and random payload
+	 *
+	 * @return void
 	 */
 	public function testError_WithDataHttpCode()
 	{
@@ -49,19 +52,19 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 		               ResponseBuilder::DEFAULT_HTTP_CODE_ERROR,
 		];
 
-		foreach($http_codes as $http_code) {
+		foreach ($http_codes as $http_code) {
 			// GIVEN data
 			$data = [$this->getRandomString('key') => $this->getRandomString('val')];
 
 			// AND error code
-			$error_code = $this->random_api_code;
+			$api_code = $this->random_api_code;
 
 			// WHEN we report error
-			$this->response = ResponseBuilder::error($error_code, null, $data, $http_code);
+			$this->response = ResponseBuilder::error($api_code, null, $data, $http_code);
 
 			// THEN returned message contains given error code and mapped message
-			$j = $this->getResponseErrorObject($error_code, $http_code);
-			$this->assertEquals($this->error_message_map[ $error_code ], $j->message);
+			$j = $this->getResponseErrorObject($api_code, $http_code);
+			$this->assertEquals($this->error_message_map[ $api_code ], $j->message);
 
 			// AND passed data
 			$this->assertEquals((object)$data, $j->data);
@@ -70,19 +73,23 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 
 	/**
 	 * Tests errorWithData()
+	 *
+	 * @return void
 	 */
 	public function testErrorWithData()
 	{
 		$data = [$this->getRandomString('key') => $this->getRandomString('val')];
-		$error_code = $this->random_api_code;
-		$this->response = ResponseBuilder::errorWithData($error_code, $data);
+		$api_code = $this->random_api_code;
+		$this->response = ResponseBuilder::errorWithData($api_code, $data);
 
-		$j = $this->getResponseErrorObject($error_code);
+		$j = $this->getResponseErrorObject($api_code);
 		$this->assertEquals((object)$data, $j->data);
 	}
 
 	/**
 	 * Tests errorWithDataAndHttpCode()
+	 *
+	 * @return void
 	 */
 	public function testErrorWithDataAndHttpCode()
 	{
@@ -92,18 +99,20 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 		               ResponseBuilder::DEFAULT_HTTP_CODE_ERROR,
 		];
 
-		foreach($http_codes as $http_code) {
+		foreach ($http_codes as $http_code) {
 			$data = [$this->getRandomString('key') => $this->getRandomString('val')];
-			$error_code = $this->random_api_code;
-			$this->response = ResponseBuilder::errorWithDataAndHttpCode($error_code, $data, $http_code);
+			$api_code = $this->random_api_code;
+			$this->response = ResponseBuilder::errorWithDataAndHttpCode($api_code, $data, $http_code);
 
-			$j = $this->getResponseErrorObject($error_code, $http_code);
+			$j = $this->getResponseErrorObject($api_code, $http_code);
 			$this->assertEquals((object)$data, $j->data);
 		}
 	}
 
 	/**
 	 * Tests errorWithDataAndHttpCode() with http_code null
+	 *
+	 * @return void
 	 *
 	 * @expectedException \InvalidArgumentException
 	 */
@@ -115,6 +124,8 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 
 	/**
 	 * Tests errorWithHttpCode()
+	 *
+	 * @return void
 	 */
 	public function testErrorWithHttpCode()
 	{
@@ -124,17 +135,19 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 		               ResponseBuilder::DEFAULT_HTTP_CODE_ERROR,
 		];
 
-		foreach($http_codes as $http_code) {
-			$error_code = $this->random_api_code;
-			$this->response = ResponseBuilder::errorWithHttpCode($error_code, $http_code);
+		foreach ($http_codes as $http_code) {
+			$api_code = $this->random_api_code;
+			$this->response = ResponseBuilder::errorWithHttpCode($api_code, $http_code);
 
-			$j = $this->getResponseErrorObject($error_code, $http_code);
+			$j = $this->getResponseErrorObject($api_code, $http_code);
 			$this->assertNull($j->data);
 		}
 	}
 
 	/**
 	 * Tests errorWithHttpCode() with @null as http_code
+	 *
+	 * @return void
 	 *
 	 * @expectedException \InvalidArgumentException
 	 */
@@ -145,29 +158,33 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 
 	/**
 	 * Tests errorWithMessageAndData()
+	 *
+	 * @return void
 	 */
 	public function testErrorWithMessageAndData()
 	{
 		$data = [$this->getRandomString('key') => $this->getRandomString('val')];
-		$error_code = $this->random_api_code;
+		$api_code = $this->random_api_code;
 		$error_message = $this->getRandomString('msg');
-		$this->response = ResponseBuilder::errorWithMessageAndData($error_code, $error_message, $data);
+		$this->response = ResponseBuilder::errorWithMessageAndData($api_code, $error_message, $data);
 
-		$j = $this->getResponseErrorObject($error_code, ResponseBuilder::DEFAULT_HTTP_CODE_ERROR, $error_message);
+		$j = $this->getResponseErrorObject($api_code, ResponseBuilder::DEFAULT_HTTP_CODE_ERROR, $error_message);
 		$this->assertEquals($error_message, $j->message);
 		$this->assertEquals((object)$data, $j->data);
 	}
 
 	/**
 	 * Tests errorWithMessage()
+	 *
+	 * @return void
 	 */
 	public function testErrorWithMessage()
 	{
-		$error_code = $this->random_api_code;
+		$api_code = $this->random_api_code;
 		$error_message = $this->getRandomString('msg');
-		$this->response = ResponseBuilder::errorWithMessage($error_code, $error_message);
+		$this->response = ResponseBuilder::errorWithMessage($api_code, $error_message);
 
-		$j = $this->getResponseErrorObject($error_code, ResponseBuilder::DEFAULT_HTTP_CODE_ERROR, $error_message);
+		$j = $this->getResponseErrorObject($api_code, ResponseBuilder::DEFAULT_HTTP_CODE_ERROR, $error_message);
 		$this->assertEquals($error_message, $j->message);
 		$this->assertNull($j->data);
 	}
@@ -175,32 +192,36 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 
 	/**
 	 * Tests error() handling api code with no message mapping
+	 *
+	 * @return void
 	 */
 	public function testError_MissingMessageMapping()
 	{
-		/** @var \MarcinOrlowski\ResponseBuilder\ErrorCode $api_codes_class_name */
+		/** @var \MarcinOrlowski\ResponseBuilder\ApiCodeBase $api_codes_class_name */
 		$api_codes_class_name = $this->getApiCodesClassName();
 
 		// FIXME we **assume** this is not mapped. But assumptions sucks...
-		$error_code = $this->max_allowed_code -1;
-		$this->response = ResponseBuilder::error($error_code);
+		$api_code = $this->max_allowed_code - 1;
+		$this->response = ResponseBuilder::error($api_code);
 
 		$key = $api_codes_class_name::getMapping($api_codes_class_name::NO_ERROR_MESSAGE);
-		$lang_args = ['error_code' => $error_code];
+		$lang_args = ['error_code' => $api_code];
 		$msg = \Lang::get($key, $lang_args);
 
-		$j = $this->getResponseErrorObject($error_code, ResponseBuilder::DEFAULT_HTTP_CODE_ERROR, $msg);
+		$j = $this->getResponseErrorObject($api_code, ResponseBuilder::DEFAULT_HTTP_CODE_ERROR, $msg);
 		$this->assertNull($j->data);
 	}
 
 	/**
 	 * Tests buildErrorResponse() fed with not allowed OK api code
 	 *
+	 * @return void
+	 *
 	 * @expectedException \InvalidArgumentException
 	 */
 	public function testBuildErrorResponse_ApiCodeOK()
 	{
-		/** @var \MarcinOrlowski\ResponseBuilder\ErrorCode $api_codes_class_name */
+		/** @var \MarcinOrlowski\ResponseBuilder\ApiCodeBase $api_codes_class_name */
 		$api_codes_class_name = $this->getApiCodesClassName();
 
 		$data = null;
@@ -214,6 +235,8 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 
 	/**
 	 * Tests buildErrorResponse() fed with api_code in form of disallowed variable type
+	 *
+	 * @return void
 	 *
 	 * @expectedException \InvalidArgumentException
 	 */
@@ -230,35 +253,39 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 	/**
 	 * Tests buildErrorResponse() fed with http_code in form of disallowed variable type
 	 *
+	 * @return void
+	 *
 	 * @expectedException \InvalidArgumentException
 	 */
 	public function testBuildErrorResponse_WrongHttpCodeType()
 	{
-		/** @var \MarcinOrlowski\ResponseBuilder\ErrorCode $api_codes_class_name */
+		/** @var \MarcinOrlowski\ResponseBuilder\ApiCodeBase $api_codes_class_name */
 		$api_codes_class_name = $this->getApiCodesClassName();
 
 		$data = null;
 		$http_code = 'string-is-invalid';
-		$error_code = $api_codes_class_name::NO_ERROR_MESSAGE;
+		$api_code = $api_codes_class_name::NO_ERROR_MESSAGE;
 		$lang_args = null;
 
-		$this->callBuildErrorResponse($data, $error_code, $http_code, $lang_args);
+		$this->callBuildErrorResponse($data, $api_code, $http_code, $lang_args);
 	}
 
 	/**
 	 * Tests buildErrorResponse() fed with @null as http_code
+	 *
+	 * @return void
 	 */
 	public function testBuildErrorResponse_NullHttpCode()
 	{
-		/** @var \MarcinOrlowski\ResponseBuilder\ErrorCode $api_codes_class_name */
+		/** @var \MarcinOrlowski\ResponseBuilder\ApiCodeBase $api_codes_class_name */
 		$api_codes_class_name = $this->getApiCodesClassName();
 
 		$data = null;
 		$http_code = null;
-		$error_code = $api_codes_class_name::NO_ERROR_MESSAGE;
+		$api_code = $api_codes_class_name::NO_ERROR_MESSAGE;
 		$lang_args = null;
 
-		$this->response = $this->callBuildErrorResponse($data, $error_code, $http_code, $lang_args);
+		$this->response = $this->callBuildErrorResponse($data, $api_code, $http_code, $lang_args);
 
 		$http_code = ResponseBuilder::DEFAULT_HTTP_CODE_ERROR;
 		$this->assertEquals($http_code, $this->response->getStatusCode());
@@ -267,56 +294,63 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 	/**
 	 * Tests buildErrorResponse() fed with http code out of allowed bounds
 	 *
+	 * @return void
+	 *
 	 * @expectedException \InvalidArgumentException
 	 */
 	public function testBuildErrorResponse_TooLowHttpCode()
 	{
-		/** @var \MarcinOrlowski\ResponseBuilder\ErrorCode $api_codes_class_name */
+		/** @var \MarcinOrlowski\ResponseBuilder\ApiCodeBase $api_codes_class_name */
 		$api_codes_class_name = $this->getApiCodesClassName();
 
 		$data = null;
 		$http_code = 0;
-		$error_code = $api_codes_class_name::NO_ERROR_MESSAGE;
+		$api_code = $api_codes_class_name::NO_ERROR_MESSAGE;
 		$lang_args = null;
 
-		$this->callBuildErrorResponse($data, $error_code, $http_code, $lang_args);
+		$this->callBuildErrorResponse($data, $api_code, $http_code, $lang_args);
 	}
 
 	/**
 	 * Tests buildErrorResponse() fed with wrong lang_args data
 	 *
+	 * @return void
+	 *
 	 * @expectedException \InvalidArgumentException
 	 */
 	public function testBuildErrorResponse_WrongLangArgs()
 	{
-		/** @var \MarcinOrlowski\ResponseBuilder\ErrorCode $api_codes_class_name */
+		/** @var \MarcinOrlowski\ResponseBuilder\ApiCodeBase $api_codes_class_name */
 		$api_codes_class_name = $this->getApiCodesClassName();
 
 		$data = null;
 		$http_code = 404;
-		$error_code = $api_codes_class_name::NO_ERROR_MESSAGE;
+		$api_code = $api_codes_class_name::NO_ERROR_MESSAGE;
 		$lang_args = 'string-is-invalid';
 
 		/** @noinspection PhpParamsInspection */
-		$this->callBuildErrorResponse($data, $error_code, $http_code, $lang_args);
+		$this->callBuildErrorResponse($data, $api_code, $http_code, $lang_args);
 	}
 
 	/**
 	 * Calls protected method buildErrorResponse()
 	 *
 	 * @param mixed|null $data
-	 * @param int|null   $error_code
+	 * @param int|null   $api_code
 	 * @param int|null   $http_code
 	 * @param mixed|null $lang_args
 	 *
 	 * @return mixed
 	 */
-	protected function callBuildErrorResponse($data, $error_code, $http_code, $lang_args)
+	protected function callBuildErrorResponse($data, $api_code, $http_code, $lang_args)
 	{
 		$obj = new ResponseBuilder();
 		$method = $this->getProtectedMethod(get_class($obj), 'buildErrorResponse');
 
-		return $method->invokeArgs($obj, [$data, $error_code, $http_code, $lang_args]);
+		return $method->invokeArgs($obj, [$data,
+		                                  $api_code,
+		                                  $http_code,
+		                                  $lang_args]);
 	}
 
 }